+++ web-rasp/src/App.css
@@ -22,6 +22,27 @@ body {
   width: 100%;
 }
 
+.search-input {
+  width: 100%;
+  padding: 0.75rem 1rem;
+  font-size: 1rem;
+  border: 2px solid #e0e0e0;
+  border-radius: 8px;
+  outline: none;
+  box-sizing: border-box;
+  background: #333;
+  cursor: pointer;
+}
+
+.search-input:focus {
+  border-color: #2196f3;
+}
+
+.date-input {
+  min-width: 150px;
+}
+
+/* ===== TABS ===== */
 .tabs {
   display: inline-flex;
   background: #f0f0f0;
@@ -52,6 +73,7 @@ body {
   box-shadow: 0 1px 3px rgba(0, 0, 0, 0.1);
 }
 
+/* ===== SEARCH CONTAINER ===== */
 .search-container {
   display: flex;
   gap: 0.75rem;
