+++ web-rasp/src/components/FavoritesList/index.js
@@ -1,0 +1,1 @@
+export { default as FavoritesList } from './FavoritesList';
