+++ web-rasp/src/components/SearchBar/SearchBar.css
@@ -5,82 +5,6 @@
   width: 100%;
 }
 
-.search-input {
-  width: 100%;
-  padding: 0.75rem 1rem;
-  font-size: 1rem;
-  border: 2px solid #e0e0e0;
-  border-radius: 8px;
-  outline: none;
-  box-sizing: border-box;
-}
-
-.search-input:focus {
-  border-color: #2196f3;
-}
-
-.search-results {
-  position: absolute;
-  top: 100%;
-  left: 0;
-  right: 0;
-  background: white;
-  border: 1px solid #e0e0e0;
-  border-radius: 8px;
-  margin-top: 0.5rem;
-  max-height: 300px;
-  overflow-y: auto;
-  list-style: none;
-  padding: 0;
-  z-index: 100;
-  box-shadow: 0 4px 12px rgba(0, 0, 0, 0.15);
-}
-
-.search-result-item {
-  padding: 0.75rem 1rem;
-  cursor: pointer;
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  border-bottom: 1px solid #f0f0f0;
-}
-
-.search-result-item:last-child {
-  border-bottom: none;
-}
-
-.search-result-item:hover {
-  background: #f5f5f5;
-}
-
-.station-title {
-  font-weight: 500;
-}
-
-.station-type {
-  font-size: 0.75rem;
-  color: #666;
-  background: #e0e0e0;
-  padding: 0.25rem 0.5rem;
-  border-radius: 4px;
-}
-
-.loading,
-.error,
-.no-results {
-  padding: 0.5rem;
-  font-size: 0.875rem;
-  width: 100%;
-}
-
-.loading {
-  color: #666;
-}
-
-.error {
-  color: #d32f2f;
-}
-
-.no-results {
-  color: #666;
+.search-bar .search-input-wrapper {
+  flex: 1;
 }
