+++ web-rasp/src/components/common/FavoriteButton.jsx
@@ -1,0 +1,13 @@
+import './FavoriteButton.css';
+
+export default function FavoriteButton({ isFavorite, onToggle, title = '' }) {
+  return (
+    <button
+      className={`favorite-button ${isFavorite ? 'active' : ''}`}
+      onClick={onToggle}
+      title={title || (isFavorite ? 'Удалить из избранного' : 'Добавить в избранное')}
+    >
+      {isFavorite ? '★' : '☆'}
+    </button>
+  );
+}
