+++ web-rasp/src/components/common/LoadingBlock.css
@@ -1,0 +1,5 @@
+.loading-block {
+  padding: 2rem;
+  text-align: center;
+  color: #666;
+}
