+++ web-rasp/src/components/common/LoadingBlock.jsx
@@ -1,0 +1,9 @@
+import './LoadingBlock.css';
+
+export default function LoadingBlock({ message = 'Загрузка...' }) {
+  return (
+    <div className="loading-block">
+      {message}
+    </div>
+  );
+}
