+++ web-rasp/src/components/common/TimeDisplay.css
@@ -1,0 +1,12 @@
+.time-display {
+  font-weight: 700;
+  color: #333;
+}
+
+.time-display.departure {
+  color: #1976d2;
+}
+
+.time-display.arrival {
+  color: #333;
+}
